(1) Default thread management in SparseX now uses the ThreadPool by default.     The ThreadPool uses a custom barrier based on the "Sense-reversing     Centralized Barrier" concept. (2) Added some user-friendly examples in the api directory. (3) Fixed the Makefile in the benchmarking framework to work according to the     supplied libraries (pOSKI, MKL). If a library is available the user must     provide the absolute address through a prefix (e.g., poski_prefix="") and     also compile with -DLIBRARY_NAME. (4) Concerning the indexing, the core works with one-based, while the api     provides both options.
Issues:
(1) The function CalcImbalance() currently doesn't work because I haven't added
    individual timers to each thread of the pool.
(2) In order to hide anything related to the pool from the api user, the pool is
    currently initialized after the tuning is complete. This can be fixed later
    on by also using the pool for the preprocessing and, thus, initialize the
    pool in the library initialization routine.

+++ csx/Vector.cpp
@@ -12,6 +12,10 @@
 #include "Vector.hpp"
 #include "Logger.hpp"
 
+#include <cstdlib>
+#include <cassert>
+#include <numa.h>
+#include <math.h>
 #include <sys/mman.h>
 
 namespace internal {
@@ -173,9 +177,9 @@ void vec_add_part(spx_vector_t *v1, spx_vector_t *v2, spx_vector_t *v3,
 		exit(1);
 	}
 
-	if (start > v1->size || end > v1->size || start > end) {
+	if ((size_t) start > v1->size || (size_t) end > v1->size || start > end) {
 		fprintf(stderr, "start=%lu end=%lu v->size=%lu not compatible\n",
-		        start, end, v1->size);
+		        (size_t) start, (size_t) end, v1->size);
 		exit(1);
 	}
